Test

Sequence

Koma.Sequence — Type

Sequence(GR)

Sequence object.

Argument

GR::Array{Grad,2} := Gradient array, fist axis is the dimension (x,y,z) and the second one is time.

Examples

julia> A = Sequence([Grad(1,1) for i=1:2,j=1:2])
Sequence(Grad[Grad(1, 1) Grad(1, 1); Grad(1, 1) Grad(1, 1)])

julia> B = Sequence([Grad(2,2) for i=1:2,j=1:2])
Sequence(Grad[Grad(2, 2) Grad(2, 2); Grad(2, 2) Grad(2, 2)])

julia> C = A + B
Sequence(Grad[Grad(1, 1) Grad(1, 1) Grad(2, 2) Grad(2, 2); Grad(1, 1) Grad(1, 1) Grad(2, 2) Grad(2, 2)])

Koma.dur — Method

Duration of the sequence T [s].

Koma.get_kspace — Method

Outputs designed k-space trajectory from sequence object.

Grad

Koma.Grad — Type

Grad_fun(f,T,N)

Generates arbitrary gradient waveform defined by function f in the interval t∈[0,T]. It uses N square gradients uniformly spaced in the interval.

Arguments

f::Function := Gradient waveform.

T::Real := Gradient duration.

N::Integer=1:= Number of sample points.

Examples

julia> Grad_fun(x-> sin(π*x),1,4)
1×4 Array{Grad,2}:
 Grad(0.0, 0.333333)  Grad(0.866025, 0.333333)  Grad(0.866025, 0.333333)  Grad(1.22465e-16, 0.333333)

Koma.Grad — Type

Grad(A,T)

Gradient object.

Arguments

A::Real := Gradient amplitude [T].

T::Real := Gradient duration [s].

Examples

Grad(1,2)

Koma.dur — Method

Duration T [s] of Grad datatype.

Koma.dur — Method

Duration T [s] of Vector{Grad}.

Koma.rotx — Method

Rotates vector counter-clockwise with respecto to the x-axis.

Koma.roty — Method

Rotates vector counter-clockwise with respecto to the y-axis.

Koma.rotz — Method

Rotates vector counter-clockwise with respecto to the z-axis.

There should be a plot here:

using Koma
grad = Grad(30e-3, 1e-3, .5e-3, 3e-3)

←3.0 ms→ Grad(30.0 mT, 1.0 ms, ↑0.5 ms, ↓0.5 ms)

seq = Sequence([grad])
p = plot_seq(seq; slider=false, height=300)
savefig(p,"assets/Grad.html"); nothing

grad2 = Grad([30,50,40].*1e-3, 1e-3, .5e-3, 1e-3)

←1.0 ms→ Grad(∿ mT, 1.0 ms, ↑0.5 ms, ↓0.5 ms)

seq = Sequence([grad;grad2;;])
p = plot_seq(seq; slider=false, height=300)
savefig(p,"assets/Grad2.html"); nothing

seq = Sequence([grad grad2 -grad -grad2 grad grad2])
p = plot_seq(seq; height=400)
savefig(p,"assets/Grad3.html"); nothing

p = plot_M0(seq; height=400)
savefig(p,"assets/Grad4.html"); nothing

RF

Koma.RF — Type

RF Object

Koma.Spinor — Type

Spinor(α,β)

Spinor(α,β) with Cayley-Klein parameters α and β.

Based on "Introduction to the Shinnar-Le Roux algorithm", Patrick Le Roux (1995).

A spinor is a way to represent 3D rotations, the underlying representation is a 2 X 2 complex unitary matrix ($\alpha,\beta\in\mathbb{C}$):

\[R=\left[\begin{array}{cc} \alpha & -\beta^{*}\\ \beta & \alpha^{*} \end{array}\right],\]

with $|\alpha|^2+|\beta|^2 = 1$.

This later operates on the $2\times2$ representation of $(x,y,z)$ as follows $V^{+} = R V R^{*}$.

Koma.Q — Method

Pauly, J., Le Roux, P., Nishimura, D., & Macovski, A. (1991). Parameter relations for the Shinnar-Le Roux selective excitation pulse design algorithm (NMR imaging). IEEE Transactions on Medical Imaging, 10(1), 53–65. doi:10.1109/42.75611

Spinor rotation matrix.

Rotation of φ with respect to the axis of rotation n=(nx,ny,nz).

\[\varphi=-\gamma\Delta t\sqrt{\left|B_{1}\right|^{2}+\left(\boldsymbol{G}\cdot\boldsymbol{x}\right)^{2}}=-\gamma\Delta t\left\Vert \boldsymbol{B}\right\Vert\]

\[\boldsymbol{n}=\boldsymbol{B}/\left\Vert \boldsymbol{B}\right\Vert \]

Koma.Rx — Method

Spinor clockwise rotation matrix with angle Θ with respect to x-axis.

Rx(θ) = Rz(-π/2) Ry(θ) Rz(π/2)

Koma.Ry — Method

Spinor clockwise rotation matrix with angle Θ with respect to y-axis.

Koma.Rz — Method

Spinor clockwise rotation matrix with angle φ with respect to z-axis.

Koma.dur — Method

Duration T [s] of the RF array Array{RF,2}.

Koma.dur — Method

Duration T [s] of RF datatype.

Koma.dur — Method

Duration T [s] of the RF array Array{RF,1}.

ADC

PulseDesigner