This page shows the documentation for the modules, structs, functions, methods and additional components available when importing the KomaMRI.jl package. It is very useful for reference when using directly the Julia REPL and when creating custom Julia scripts. Please, don't miss out the section How to read the API docs which has important considerations to understand general aspects of the docstring structure. The following are the contents of the API Documentation:

How to read the API docs

Structs

Scanner

Phantom

Sequence

Grad

RF

ADC

Delay

Read Data

read_seq

read_phantom_jemris

signal_to_raw_data

Pulse Design

PulseDesigner

PulseDesigner.RF_hard

PulseDesigner.EPI

PulseDesigner.radial_base

Simulation

simulate

simulate_slice_profile

Plots

plot_phantom_map

plot_seq

plot_kspace

plot_M0

plot_signal

plot_image

How to read the API docs

The API documentation has predefined "template patterns" which are meant to help the user to understand how to use modules, structs, functions, methods and every aspect necessary in order to take advantage of all the possibilities that KomaMRI.jl offers.

These documentation "template patterns" are based from the Julia Blue Style documentation and other github repositories that works with MRI topics. However, some custom considerations were added for a better understanding and a wider perspective.

Whenever you see a docstring documentation, it will have the following structure:

KomaMRI.component_name — Component

out1, out2, ... = component_name(arg1, arg2, ...; kw1, kw2, ...)

This is a brief description of what component_name does.

Note

Here can be placed a note if it is regarded necessary.

Arguments

arg1: (::type, =value, [unit], opts: [opt1, opt2, ...]) the description for the arg1

...

Keywords

kw1: (::type, =value, [unit], opts: [opt1, opt2, ...]) the description for the kw1

...

Returns

out1: (::type, =value, [unit], opts: [opt1, opt2, ...]) the description for the out1

...

References

Sometimes it is a good idea to put some references or links

...

Examples

julia> arg1, arg2, valkw1, valkw2 = 3.5, "hello", 1, true

julia> out1, out2 = component_name(arg1, arg2; kw1=valkw1, kw2=valkw2)

The previous docstring block will always have first the way how it is called the component (outputs = component_name(inputs) and next a brief description about what the component does. Then a note block will be displayed if necessary. As general rule, the next subsections are optional: Arguments, Keywords, Returns, References and Examples, however they will be displayed whenever necessary. These subsections are self-explanatory, so it is intuitive to figure out what are they meant for.

Note that every subitem in the sections Arguments, Keywords and Returns are variables. They have practical information enclosed in parentheses plus a description. They information in parentheses is optional but very useful if it is present:

::type: is the suggested type of the variable. If the input variable is of type ::type, then nothing can go wrong, but it is always possible to test other types. If the variable is an output, then it will always try to be forced to the type ::type.

=value: sometimes for the inputs the is defined a default value if it is not assigned by the user.

[unit]: this is the suggested physical unit of measure of the variable. Everything it is going to be fine if you are stick with these units of measure.

opts: [opt1, opt2, ...]: sometimes the input value can only be interpreted if it is part of some predefined values.

Structs

Scanner

KomaMRI.Scanner — Type

scanner = Scanner(B0, B1, Gmax, Smax, ADC_Δt, seq_Δt, GR_Δt, RF_Δt,
    RF_ring_down_T, RF_dead_time_T, ADC_dead_time_T)

The Scanner struct.

Arguments

B0: (::Real, =1.5, [T]) main magnetic field

B1: (::Real, =10e-6, [T]) maximum RF amplitude

Gmax: (::Real, =60e-3, [T/m]) maximum Gradient

Smax: (::Real, =500, [mT/m/ms]) maximum slew-rate

ADC_Δt: (::Real, =2e-6, [s]) ADC raster time

seq_Δt: (::Real, =1e-5, [s]) sequence-block raster time

GR_Δt: (::Real, =1e-5, [s]) gradient raster time

RF_Δt: (::Real, =1e-6, [s]) RF raster time

RF_ring_down_T: (::Real, =20e-6, [s]) RF ring down time

RF_dead_time_T: (::Real, =100e-6, [s]) RF dead time

ADC_dead_time_T: (::Real, =10e-6, [s]) ADC dead time

Returns

scanner: (::Scanner) Scanner struct

Phantom

KomaMRI.Phantom — Type

phantom = Phantom(name, x, y, z, ρ, T1, T2, T2s, Δw, Dλ1, Dλ2, Dθ, ux, uy, uz)

The Phantom struct.

Arguments

name: (::String) name of the Phantom

x: (::Vector{Float64}, [m]) vector of x-positions of the spins

y: (::Vector{Float64}, [m]) vector of y-positions of the spins

z: (::Vector{Float64}, [m]) vector of z-positions of the spins

ρ: (::Vector{Float64}) vector of proton density of the spins

T1: (::Vector{Float64}, [s]) vector of T1 parameters of the spins

T2: (::Vector{Float64}, [s]) vector of T2 parameters of the spins

T2s: (::Vector{Float64}, [s]) vector of T2s parameters of the spins

Δw: (::Vector{Float64}, [rad/s]) vector of off-resonance parameters of the spins

Dλ1: (::Vector{Float64}) vector of Dλ1 (diffusion) parameters of the spins

Dλ2: (::Vector{Float64}) vector of Dλ2 (diffusion) parameters of the spins

Dθ: (::Vector{Float64}) vector of Dθ (diffusion) parameters of the spins

ux: (::Function) displacement field in the x-axis

uy: (::Function) displacement field in the y-axis

uz: (::Function) displacement field in the z-axis

Returns

phantom: (::Phantom) Phantom struct

KomaMRI.brain_phantom2D — Function

phantom = brain_phantom2D(;axis="axial", ss=4)

Creates a two-dimentional brain phantom struct.

References

B. Aubert-Broche, D.L. Collins, A.C. Evans: "A new improved version of the realistic digital brain phantom" NeuroImage, in review - 2006

B. Aubert-Broche, M. Griffin, G.B. Pike, A.C. Evans and D.L. Collins: "20 new digital brain phantoms for creation of validation image data bases" IEEE TMI, in review - 2006

https://brainweb.bic.mni.mcgill.ca/brainweb

Keywords

axis: (::String, ="axial", opts=["axial"]) orientation of the phantom

ss: (::Real, =4) subsampling parameter in all axis

Returns

phantom: (::Phantom) 2D Phantom struct

Examples

julia> obj = brain_phantom2D()

julia> plot_phantom_map(obj, :ρ)

KomaMRI.brain_phantom3D — Function

phantom = brain_phantom3D(;ss=4)

Creates a three-dimentional brain phantom struct.

References

B. Aubert-Broche, D.L. Collins, A.C. Evans: "A new improved version of the realistic digital brain phantom" NeuroImage, in review - 2006

B. Aubert-Broche, M. Griffin, G.B. Pike, A.C. Evans and D.L. Collins: "20 new digital brain phantoms for creation of validation image data bases" IEEE TMI, in review - 2006

https://brainweb.bic.mni.mcgill.ca/brainweb

Keywords

ss: (::Real, =4) subsampling parameter in all axis

Returns

phantom: (::Phantom) 3D Phantom struct

Examples

julia> obj = brain_phantom3D()

julia> plot_phantom_map(obj, :ρ)

Sequence

KomaMRI.Sequence — Type

seq = Sequence()
seq = Sequence(GR)
seq = Sequence(GR, RF)
seq = Sequence(GR, RF, ADC)
seq = Sequence(GR, RF, ADC, DUR)
seq = Sequence(GR::Array{Grad,1})
seq = Sequence(GR::Array{Grad,1}, RF::Array{RF,1})
seq = Sequence(GR::Array{Grad,1}, RF::Array{RF,1}, A::ADC, DUR, DEF)

The Sequence struct.

Arguments

GR: (::Matrix{Grad}) gradient matrix, rows are for (x,y,z) and columns are for time

RF: (::Matrix{RF}) RF matrix, the 1 row is for the coil and columns are for time

ADC: (::Vector{ADC}) ADC vector in time

DUR: (::Vector{Float64}, [s]) duration of each sequence-block, this enables delays after RF pulses to satisfy ring-down times

DEF: (::Dict{String, Any}) dictionary with relevant information of the sequence. The possible keys are ["AdcRasterTime", "GradientRasterTime", "Name", "Nz", "Num_Blocks", "Nx", "Ny", "PulseqVersion", "BlockDurationRaster", "FileName", "RadiofrequencyRasterTime"]

Returns

seq: (::Sequence) Sequence struct

Grad

KomaMRI.Grad — Type

grad = Grad(A, T)
grad = Grad(A, T, rise)
grad = Grad(A, T, rise, delay)
grad = Grad(A, T, rise, fall, delay)

The Gradient struct.

Arguments

A: (::Float64, [T]) amplitude of the gradient

T: (::Float64, [s]) duration of the flat-top

rise: (::Real, [s]) duration of the rise

fall: (::Real, [s]) duration of the fall

delay: (::Real, [s]) duration of the delay

Returns

grad: (::Grad) gradient struct

KomaMRI.Grad — Method

grad = Grad(f::Function, T::Real, N::Int64; delay::Real)

Generates an arbitrary gradient waveform defined by function f in the interval t ∈ [0,T]. It uses N square gradients uniformly spaced in the interval.

Arguments

f: (::Function) function that describes the gradient waveform

T: (::Real, [s]) duration of the gradient waveform

N: (::Int64) number of samples of the gradient waveform

Keywords

delay: (::Real, =0, [s]) starting delay for the waveform

Returns

grad: (::Grad) gradient struct

Examples

julia> f1 = t -> sin(π*t / 0.8)

julia> seq = Sequence([Grad(f1, 0.8)])

julia> plot_seq(seq)

RF

KomaMRI.RF — Type

rf = RF(A, T)
rf = RF(A, T, Δf)
rf = RF(A, T, Δf, delay)

The RF struct.

Arguments

A: (::Complex{Int64}, [T]) the amplitud-phase B1x + i B1y

T: (::Int64, [s]) the duration of the RF

Δf: (::Float64, [Hz]) the frequency offset of the RF

delay: (::Float64, [s]) the delay time of the RF

Returns

rf: (::RF) the RF struct

ADC

KomaMRI.ADC — Type

adc = ADC(N, T)
adc = ADC(N, T, delay)
adc = ADC(N, T, delay, Δf, ϕ)

The ADC struct.

Arguments

N: (::Int64) number of acquired samples

T: (::Float64, [s]) duration to acquire the samples

delay: (::Float64, [s]) delay time to start the acquisition

Δf: (::Float64, [Hz]) delta frequency. It's meant to compensate RF pulse phases. It is used internally by the read_ADC function

ϕ: (::Float64, [rad]) phase. It's meant to compensate RF pulse phases. It is used internally by the read_ADC function

Returns

adc: (::ADC) ADC struct

Delay

KomaMRI.Delay — Type

delay = Delay(T)

The Delay struct. The input delay time T must be non-negative.

Note

This struct is meant to add delays to a sequence struct.

Arguments

T: (::Real, [s]) time delay value

Returns

delay: (::Delay) delay struct

Read Data

read_seq

KomaMRI.read_seq — Function

seq = read_seq(filename)

Returns the Sequence struct from a sequence file .seq.

Arguments

filename: (::String) the absolute or relative path of the sequence file .seq

Returns

seq: (::Sequence) Sequence struct

Examples

julia> seq = read_seq("examples/1.sequences/spiral.seq")

julia> plot_seq(seq)

read_phantom_jemris

KomaMRI.read_phantom_jemris — Function

phantom = read_phantom_jemris(filename)

Returns the Phantom struct from a JEMRIS phantom file .h5.

Arguments

filename: (::String) the absolute or relative path of the phantom file .h5

Returns

phantom: (::Phantom) Phantom struct

Examples

julia> obj = read_phantom_jemris("examples/2.phantoms/brain.h5")

julia> plot_phantom_map(obj, :ρ)

signal_to_raw_data

KomaMRI.signal_to_raw_data — Function

raw_ismrmrd = signal_to_raw_data(signal, seq; phantom, sys, simParams)

Transforms the raw signal into ISMRMRD format.

Arguments

signal: (::Vector{ComplexF64}) raw signal

seq: (::Sequence) Sequence struct

Keywords

phantom: (::Phantom, =Phantom(name="Phantom",x=[0])) Phantom struct

sys: (::Scanner, =Scanner()) Scanner struct

simParams: (::Dict{String,Any}(), =Dict{String,Any}()) dictionary with simulation parameters

Returns

raw_ismrmrd: (::RawAcquisitionData) raw signal in ISMRMRD format

Examples

julia> signal = simulate(obj, seq, sys; simParams=Dict{String,Any}("return_type"=>"mat"));

julia> ismrmrd = signal_to_raw_data(signal, seq; phantom=obj, sys=sys);

julia> plot_signal(ismrmrd)

Pulse Design

PulseDesigner

KomaMRI.PulseDesigner — Module

PulseDesigner

A module to define different pulse sequences.

PulseDesigner.RF_hard

KomaMRI.PulseDesigner.RF_hard — Function

ex = RF_hard(B1, T, sys::Scanner; G=[0,0,0], Δf=0)

Definition of the RF hard sequence.

Arguments

B1: (Float64, [T]) amplitude of the RF pulse

T: (Float64, [s]) duration of the RF pulse

sys: (::Scanner) Scanner struct

Keywords

G: (Vector{Float64}, =[0, 0, 0], [T]) gradient amplitudes for x, y, z

Δf: (Float64, =0, [Hz]) frequency offset of the RF pulse

Returns

ex: (::Sequence) excitation Sequence struct

Examples

julia> sys = Scanner();

julia> durRF = π/2/(2π*γ*sys.B1); #90-degree hard excitation pulse

julia> ex = PulseDesigner.RF_hard(sys.B1, durRF, sys)
Sequence[ τ = 0.587 ms | blocks: 1 | ADC: 0 | GR: 0 | RF: 1 | DEF: 0 ]

julia> plot_seq(ex)

PulseDesigner.EPI

KomaMRI.PulseDesigner.EPI — Function

epi = EPI(FOV::Float64, N::Int, sys::Scanner)

Definition of the EPI sequence.

Arguments

FOV: (::Float64, [m]) field of view

N: (::Int) number of pixels in the x and y axis

sys: (::Scanner) Scanner struct

Returns

epi: (::Sequence) epi Sequence struct

Examples

julia> sys = Scanner();

julia> FOV, N = 23e-2, 101;

julia> epi = PulseDesigner.EPI(FOV, N, sys)
Sequence[ τ = 62.259 ms | blocks: 203 | ADC: 101 | GR: 205 | RF: 0 | DEF: 4 ]

julia> plot_seq(epi)

PulseDesigner.radial_base

KomaMRI.PulseDesigner.radial_base — Function

seq = radial_base(FOV::Float64, Nr::Int, sys::Scanner)

Definition of the radial base sequence.

Arguments

FOV: (::Float64, [m]) field of view

N: (::Int) number of pixels along the radious

sys: (::Scanner) Scanner struct

Returns

seq: (::Sequence) radial base Sequence struct

Simulation

simulate

KomaMRI.simulate — Function

out = simulate(obj::Phantom, seq::Sequence, sys::Scanner; simParams, w)

Returns the raw signal or the last state of the magnetization according to the value of the "return_type" key of the simParams dictionary.

Arguments

obj: (::Phantom) Phantom struct

seq: (::Sequence) Sequence struct

sys: (::Scanner) Scanner struct

Keywords

simParams: (::Dict{String,Any}, =Dict{String,Any}()) the dictionary with simulation parameters

w: (::Any, =nothing) the flag to regard a progress bar in the blink window UI. If this variable is differnet from nothing, then the progress bar is considered

Returns

out: (::Vector{ComplexF64} or ::S <: SpinStateRepresentation or RawAcquisitionData) depending if "return_type" is "mat" or "state" or "raw" (default) respectively.

Examples

julia> sys, obj, seq = Scanner(), brain_phantom2D(), read_seq("examples/1.sequences/spiral.seq");

julia> ismrmrd = simulate(obj, seq, sys);

julia> plot_signal(ismrmrd)

simulate_slice_profile

KomaMRI.simulate_slice_profile — Function

M = simulate_slice_profile(seq; z, simParams)

Returns magnetization of spins distributed along z after running the Sequence seq.

Arguments

seq: (::Sequence) Sequence struct

Keywords

z: (=range(-2e-2,2e-2,200)) range for the z axis

simParams: (::Dict{String, Any}, =Dict{String,Any}("Δt_rf"=>1e-6)) dictionary with simulation parameters

Returns

M: (::Vector{Mag}) final state of the Mag vector

Plots

plot_phantom_map

KomaMRI.plot_phantom_map — Function

p = plot_phantom_map(ph, key; t0=0, height=600, width=nothing, darkmode=false)

Plots a phantom map for a specific spin parameter given by key.

Arguments

ph: (::Phantom) Phantom struct

key: (::Symbol, opts: [:ρ, :T1, :T2, :T2s, :x, :y, :z]) symbol for displaying different parameters of the phantom spins

Keywords

t0: (::Float64, =0, [ms]) time to see displacement of the phantom

height: (::Int64, =600) height of the plot

width: (::Int64, =nothing) width of the plot

darkmode: (::Bool, =false) boolean to define colors for darkmode

Returns

p: (::PlotlyJS.SyncPlot) plot of the phantom map for a specific spin parameter

References

Colormaps from https://github.com/markgriswold/MRFColormaps Towards Unified Colormaps for Quantitative MRF Data, Mark Griswold, et al. (2018).

Examples

julia> obj2D, obj3D = brain_phantom2D(), brain_phantom3D();

julia> plot_phantom_map(obj2D, :ρ)

julia> plot_phantom_map(obj3D, :ρ)

plot_seq

KomaMRI.plot_seq — Function

p = plot_seq(seq; width, height, slider, show_seq_blocks, show_sim_blocks, Nblocks,
        darkmode, max_rf_samples, range)

Plots a sequence struct.

Arguments

seq: (::Sequence) Sequence struct

Keywords

width: (::Int64, =nothing) width of the plot

height: (::Int64, =nothing) height of the plot

slider: (::Bool, =true) boolean to display a slider

show_seq_blocks: (::Bool, =false) boolean to show sequence blocks

show_sim_blocks: (::Bool, =false) boolean to show simulation blocks

Nblocks: (::Int64, =0) number of simulation blocks to display

darkmode: (::Bool, =false) boolean to define colors for darkmode

max_rf_samples: (::Int64, =100) maximum number of RF samples

range: (::Vector{Float64}, =[]) time range to be displayed initially

Returns

p: (::PlotlyJS.SyncPlot) plot of the Sequence struct

Examples

julia> seq = read_seq("examples/1.sequences/spiral.seq")

julia> plot_seq(seq)

plot_kspace

KomaMRI.plot_kspace — Function

p = plot_kspace(seq; width=nothing, height=nothing, darkmode=false)

Plots the k-space of a sequence struct.

Arguments

seq: (::Sequence) Sequence struct

Keywords

width: (::Int64, =nothing) width of the plot

height: (::Int64, =nothing) height of the plot

darkmode: (::Bool, =false) boolean to define colors for darkmode

Returns

p: (::PlotlyJS.SyncPlot) plot of the k-space of the sequence struct seq

Examples

julia> seq = read_seq("examples/1.sequences/spiral.seq")

julia> plot_kspace(seq)

plot_M0

KomaMRI.plot_M0 — Function

p = plot_M0(seq; height=nothing, width=nothing, slider=true, darkmode=false)

Plots the magnetization M0 of a sequence struct.

Arguments

seq: (::Sequence) Sequence struct

Keywords

height: (::Int64, =nothing) height of the plot

width: (::Int64, =nothing) width of the plot

slider: (::Bool, =true) boolean to display a slider

darkmode: (::Bool, =false) boolean to define colors for darkmode

Returns

p: (::PlotlyJS.SyncPlot) plot of the magnetization M0 of the sequence struct seq

Examples

julia> seq = read_seq("examples/1.sequences/spiral.seq")

julia> plot_M0(seq)

plot_signal

KomaMRI.plot_signal — Function

p = plot_signal(raw::RawAcquisitionData; height, width, slider, show_sim_blocks,
        darkmode, range)

Plots a raw signal in ISMRMRD format.

Arguments

raw: (::RawAcquisitionData) RawAcquisitionData struct which is the raw signal in ISMRMRD format

Keywords

width: (::Int64, =nothing) width of the plot

height: (::Int64, =nothing) height of the plot

slider: (::Bool, =true) boolean to display a slider

show_sim_blocks: (::Bool, =false) boolean to show simulation blocks

darkmode: (::Bool, =false) boolean to define colors for darkmode

range: (::Vector{Float64}, =[]) time range to be displayed initially

Returns

p: (::PlotlyJS.SyncPlot) plot of the raw signal

Examples

julia> sys, obj, seq = Scanner(), brain_phantom2D(), read_seq("examples/1.sequences/spiral.seq");

julia> ismrmrd = simulate(obj, seq, sys);

julia> plot_signal(ismrmrd)

plot_image

KomaMRI.plot_image — Function

p = plot_image(image; height, width, zmin, zmax, darkmode, title)

Plots an image matrix.

Arguments

image: (::Matrix{Float64}) image matrix

Keywords

height: (::Int64, =750) height of the plot

width: (::Int64, =nothing) width of the plot

zmin: (::Float64, =minimum(abs.(image[:]))) reference value for minimum color

zmax: (::Float64, =maximum(abs.(image[:]))) reference value for maximum color

darkmode: (::Bool, =false) boolean to define colors for darkmode

title: (::String, ="") title of the plot

Returns

p: (::PlotlyJS.SyncPlot) plot of the image matrix